further componentize product picker
* break out browse list/detail rendering into dedicated component
* break out search list/detail rendering into a dedicated component
