+++ web/js/components/layer/product-picker/layer-list.js
@@ -80,14 +80,14 @@ class LayerList extends React.Component {
       measurementConfig,
       layerConfig,
       addLayer,
-      category,
       removeLayer,
-      selectedProjection,
       activeLayers,
-      hasMeasurementSource,
+      selectedProjection,
       selectedMeasurement,
+      hasMeasurementSource,
       hasMeasurementSetting,
       updateSelectedMeasurement,
+      category,
       categoryConfig,
       setSourceIndex,
       selectedDate,
